app/pyproject.toml
[tool.poetry]
name = "Western-Friend-website"
version = "0.1.0"
description = "Website for Western Friend (westernfriend.org)"
authors = ["Brylie Christopher Oxley <brylie@amble.fi>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
arrow = "^0.15.4"
braintree = "^3.57.1"
django-crispy-forms = "^1.8.1"
wagtailfontawesome = "^1.1.4"
django-registration = "^3.0.1"
django-flatpickr = "^1.0.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.4"
pandas = "^1.0.3"
python-dotenv = "^0.14.0"
tqdm = "^4.57.0"
django-timezone-field = "^5.0"
tzdata = "^2022.1"
wagtail-color-panel = "^1.3.1"
django-storages = "^1.13.1"
boto3 = "^1.24.46"
wagtailmedia = "^0.12.0"
wagtail = "^4.1.1"
django-extensions = "^3.2.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
autopep8 = "^1.4.4"
pylint = "^2.5.3"
black = "^22.3.0"

[tool.poetry.group.dev.dependencies]
autohooks = "^22.11.2"
autohooks-plugin-black = "^22.11.0"
autohooks-plugin-isort = "^22.8.0"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"


[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.isort"]

[tool.autohooks.plugins.black]
arguments = ["-q", "--diff"]

[tool.isort]
profile = "black"

[tool.ruff]
# Same as Black.
line-length = 88

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "**/migrations"
]

# Assume Python 3.10.
target-version = "py310"
